openvpn-example: Move OpenVPN article to my blog
As I noted back in f1f56a7ae6, I originally wrote the article on
assigning OpenVPN instances to network namespaces as a blog post. So now
that I've finally set up a blog and posted the article there, remove the
duplicate copy in this repository and provide links to the new post.

Honestly, because the article can stand on its own as a guide (i.e., it
doesn't require nsdo), it never belonged in the nsdo repository in the
first place. The only benefit I see is that GitHub has a much higher
PageRank than my blog does, meaning my blog will probably show up much
later (if at all) on Google search results. But oh well.